requirements.txt
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.29.0
SQLAlchemy==2.0.30
aiomysql==0.2.0
aiosqlite==0.20.0
aiofiles==23.2.1
pydantic==2.7.1
pillow==10.3.0
pillow-avif-plugin==1.4.3
moviepy==1.0.3
opencv-python==4.9.0.80